+++ src/Command/UpdateCommand.php
@@ -18,6 +18,7 @@ use DateTimeInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
 
 class UpdateCommand extends Command {
 
@@ -28,27 +29,38 @@ class UpdateCommand extends Command {
   protected static $defaultDescription = 'Remove completed tasks from all project documents';
 
   /**
+   * @var \Symfony\Component\Console\Input\InputInterface
+   */
+  private InputInterface $input;
+
+  /**
    * @var \Symfony\Component\Console\Output\OutputInterface
    */
   private OutputInterface $output;
 
   private string $directioDirectory;
 
+  private SymfonyStyle $io;
+
+  private function io(): SymfonyStyle {
+    return $this->io ??= new SymfonyStyle($this->input, $this->output);
+  }
+
   protected function configure() {
   }
 
   protected function execute(InputInterface $input, OutputInterface $output) {
+    $this->input = $input;
+    $this->output = $output;
     $base_dir = $this->getBaseDirOrInitializeCurrent($input, $output);
     if (!$base_dir) {
       return Command::FAILURE;
     }
     $this->directioDirectory = $base_dir . DIRECTORY_SEPARATOR . Names::FILENAME_INIT;
     $now = date_create('now', LocalTimezone::get());
-
-    $this->output = $output;
     if (!file_exists($this->directioDirectory)) {
-      $output->writeln('<error>Current directory is not initialized.</error>');
-      $output->writeln(sprintf('<info>Try the "%s" command first.</info>', InitializeCommand::getDefaultName()));
+      $this->io()->error('Current directory is not initialized.');
+      $this->io()->info(sprintf('Try the "%s" command first.', InitializeCommand::getDefaultName()));
 
       return Command::FAILURE;
     }
@@ -56,8 +68,8 @@ class UpdateCommand extends Command {
     $files_to_update = glob($this->directioDirectory . DIRECTORY_SEPARATOR . Names::FILENAME_IMPORTED . DIRECTORY_SEPARATOR . '*');
     $shortpath_directio = (new GetShortPath())($this->directioDirectory);
     if (empty($files_to_update)) {
-      $output->writeln(sprintf('<error>No documents in "%s"</error>', $shortpath_directio));
-      $output->writeln(sprintf('<info>Try the "%s" command first.</info>', ImportCommand::getDefaultName()));
+      $this->io()->error(sprintf('No documents in "%s"', $shortpath_directio));
+      $this->io()->info(sprintf('Try the "%s" command first.', ImportCommand::getDefaultName()));
 
       return Command::FAILURE;
     }
@@ -67,7 +79,7 @@ class UpdateCommand extends Command {
 
     foreach ($files_to_update as $document_path) {
       $document_label = (new GetShortPath())($document_path);
-      $output->writeln($document_label);
+      $this->io()->writeln($document_label);
       $document = (new ReadDocument())($document_path);
       (new ValidateTaskSyntax())($document->getContent());
       $completed_tasks = (new ScanForCompletedTasks())($document->getContent());
@@ -78,7 +90,7 @@ class UpdateCommand extends Command {
 
       foreach ($completed_tasks as $task_data) {
         $task_id = SpecialAttributes::extractId($task_data);
-        $output->writeln('✅ ' . $task_id);
+        $this->io()->writeln('✅ ' . $task_id);
         $document = $document->withoutTask($task_id);
 
         $task = (new TaskState())